third_party/blink/web_tests/fast/events/mutation-during-replace-child-expected.txt

Ensures that replaceChild() throws an exception if mutation even handler does something wrong

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS target.replaceChild(newChild, oldChild); threw exception NotFoundError: Failed to execute 'replaceChild' on 'Node': The node to be removed is no longer a child of this node. Perhaps it was moved in response to a mutation?.
PASS successfullyParsed is true

TEST COMPLETE
